`StateMachine`: fix a test failure where an injected error is swallowed
The test failure can be reproduced using the following seed:

```
cabal run lsm-tree-test -- \
  -p '$0=="lsm-tree.Test.Database.LSMTree.StateMachine.propLockstep_RealImpl_MockFS_IO"' \
  --quickcheck-replay="(SMGen 11798857835899383018 1561550946174790845,97)"
```

The test fails because an injected error with an `FsReachedEOF` type is
swallowed by the `openTableFromSnapshot` public API function. We fix the test
failure (for now) by not generating errors of that type.

At some point we should think of a more principled approach to this type of
swallowed error. `FsReachedEOF` is not an error type that any of the `HasFS`
primitives can throw in practice when using the real file system, but `fs-sim`
error injection *can* throw this error type. Maybe `fs-sim` shouldn't inject
this type of error.