Move `IsList` and `IsString` instances for `RawBytes` to utils
I think their main use case is as utilities for testing -- using them in
production code would probably be bad for performance, and there are probably
not a lot of use cases for it.

I also think the `Semigroup` and `Monoid` instances are iffy, but it's too much
of hassle to remove those now.