Make the MergeBatchSize an adjustable parameter in TableConfig
Previously it was hard coded to be the same as the write buffer size.

Document what it means as a new tunable parameter.

Setting this low (1) is important for getting good parallel work balance
on the pipelined WP8 benchmark. It is a crucial change that makes the
pipelined version actually improve performance. Previously it would only
get about a 5 to 10% improvement.