Issue #755: fix the bug in `maxPhysicalDebt` and `invariant`
The prototype had a bug in the calculation of `maxPhysicalDebt` where it was not
considering that incoming runs that are inputs to a merge can be overfull. Runs
can be overfull because we hold back runs that are underfull. However, runs
should only be overfull in the sense that they are too large for their current
level. They can't be overfull with respect to the next level too.

This same bug occurred in `invariant`, where a similar change was made to
consider overfull runs properly.

Some other minor changes to `invariant` were made to make the checked invariant
a little more precise.