Introduce a V1 snapshot serialisation format version
But don't yet actually change the serialisation format. This is partly
just to demonstrate to ourselves how to do it, so there's a pattern to
follow in future.

Doing this highlights that we cannot generally match on the version, and
should only do so in places where the format is actually different
between versions. Otherwise we would have to duplicate too much code.