Prefix snapshot codec golden files using the current snapshot version
As a result, we keep around golden files for old snapshot versions as newer ones
are created. The upside of this is that we could test the backwards compatiblity
of the snapshot codec: run versioned decoders on the golden files for older
snapshot versions and check that there are no errors. I've not implemented such
a test yet, but this would be one step in that direction.