`lsm-tree`: make the hash salt for bloom filters configurable
Making the hash salt configurable improves security, because bloom filter hashes
are not cryptographic hashes.

The hash salt can be configured on a session-wide basis only. That is, all bloom
filters for all tables in a single session use the same salt. As a result, batch
lookup performance is not impacted by the salt. The performance currently relies
on being able to compute a hash only once for multiple bloom filters, which is
only possible if the salt for each bloom filter is the same.

For now the user has the responsibility of passing in the same salt each time a
session is restored. We will change this in the next few commits.

Co-authored-by: Wen Kokke <wenkokke@users.noreply.github.com>