Split `openSession` into `newSession` and `restoreSession`
Only the former requires a salt, and the latter does not require it. We still
keep a version of `openSession` that defers to `newSession` or `restoreSession`
based on the session directory contents, because we expect it to be useful for
users.

In `newSession`, the input salt is written to a new metadata file in the session
directory. In `restoreSession`, we read this metadata file back into memory to
extract the salt.