Add a script that can be used to check release builds
Basically, if a release tag of the form `blockio-*`, `bloomfilter-blocked-*` or
`lsm-tree-*` is created, the workflow will trigger and it will try to build the
corresponding library without the local dev environment. Instead, it will build
the library without tests and benchmark with the latest Hackage index-state.
This should be a sanity check that the library version will work out of the box
once it is released to Hackage