Limit the scope of forgotten reference checks to sessions
We do this to check for forgotten references in isolation from other parts of a
program. For example, if we have two separate properties that we want to run in
the same test suite, then we don't want a forgotten reference from one property
show up in the other property.