`blockio`: guard that buffers are pinned in `submitIO`
The `Async` implementation was already doing this, but now the `IOError` is
thrown as an `FsError`. The `Serial` implementation now also throws an error
when buffers are unpinned, which was done to make behaviour across
implementations more uniform. I have also added tests to check that "unpinned
buffer" errors are indeed being thrown.