`utxo-bench`: introduce a new `RunMode` option
Previously, the `run` command of the benchmark executable had two flags:
`--pipelined` and `--lookup-only`. The value of the second flag had no effect if
`--pipelined` was passed to the executable. Now there is a new `--mode` option
that can be set to `Sequential`, `Pipelined`, or `LookupOnly`. `Sequential` is
the name for what previously was the default mode of the `run` command.