Split out a GHA workflow file for documentation jobs
Due to the machinations of `cabal haddock-project`, caching has to be done
differently when we build the dev haddocks compared to when we build and test
the project. It seems simpler to put documentation jobs in their own workflow
and update the caching approach there, though that does mean there is some
duplication between the documentation workflow and the build-and-test workflow.